\name{GGdata}
\alias{GGdata}
\docType{data}
\title{
SNP Genotypes on \emph{Miscanthus sinensis}
}
\description{
GoldenGate SNP genotypes for 581 individuals and 395 markers.  The numbers
zero and two represent homozygotes, and one represents heterozygotes.
}
\usage{data("GGdata")}
\format{
The data are stored in a numeric matrix with individuals in rows and markers
in columns.
}

\source{
\url{http://hdl.handle.net/2142/49963}

Clark et al. (2014) A footprint of past climate change on the diversity and 
population structure of \emph{Miscanthus sinensis}. Annals of Botany 114(1):97-107.
}

\examples{
data(GGdata)
}
\keyword{datasets}
